requirements.txt
requests
httpx
supabase
python-geohash
discordwebhook
tenacity
pytz